requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==1.4.52
passlib[bcrypt]==1.7.4
python-jose==3.3.0
aiofiles==23.1.0
python-dotenv==1.0.0
